requirements.txt
streamlit==1.29.0
python-dotenv==1.0.1

pandas==2.1.4
numpy==1.26.4

plotly==5.18.0
matplotlib==3.8.4
seaborn==0.13.2

yfinance==0.2.36
ta==0.11.0
statsmodels==0.14.1

scikit-learn==1.4.2
xgboost==2.0.3
lightgbm>=4.0.0
prophet>=1.1.0

# TensorFlow is optional - install manually if needed for LSTM
# tensorflow>=2.15.0

vaderSentiment==3.3.2

pytz==2024.1
